모델 수정 (2)
