+++ kurlyClone/Models/ApiResult/AuthApiResult.swift
@@ -1,0 +1,8 @@
+//
+//  AuthApiResult.swift
+//  kurlyClone
+//
+//  Created by 고성민 on 2023/08/25.
+//
+
+import Foundation
